Controller and control command interface separated
