+++ src/beamlime/applications/controller.py
@@ -3,49 +3,40 @@
 
 from __future__ import annotations
 
-from typing import NewType
-
-from ..empty_factory import empty_app_factory
-from ..logging import BeamlimeLogger
-from ..logging.mixins import LogMixin
-
-StartedFlag = NewType("StartedFlag", bool)
-default_started = StartedFlag(False)
-
-PausedFlag = NewType("PausedFlag", bool)
-default_paused = PausedFlag(False)
+from ..applications.interfaces import BeamlimeApplicationInterface, ControlInterface
+from ..empty_factory import empty_app_factory as app_factory
+from .data_generator import DataGenerator
+from .data_reduction import DataFeeder
+from .data_visualization import DataPlotter
 
 
-class ApplicationNotStartedError(Exception):
-    ...
+@app_factory.provider
+class Controller(BeamlimeApplicationInterface):
+    remote_ctrl: ControlInterface
+    data_generator: DataGenerator
+    data_feeder: DataFeeder
+    data_plotter: DataPlotter
 
+    def run(self):
+        from asyncio import create_task, get_running_loop
 
-class ApplicationNotResumedError(Exception):
-    ...
+        self.remote_ctrl.start()
+        try:
+            event_loop = get_running_loop()
+        except RuntimeError:
+            from asyncio import run, wait
 
+            async def _run():
+                await wait(
+                    (
+                        create_task(self.data_feeder.run()),
+                        create_task(self.data_generator.run()),
+                        create_task(self.data_plotter.run()),
+                    )
+                )
 
-class ControlInterface(LogMixin):
-    """Command Board singleon that all applications will share."""
-
-    logger: BeamlimeLogger
-    started: StartedFlag = default_started
-    paused: PausedFlag = default_paused
-
-    def start(self):
-        self.started = StartedFlag(True)
-        self.debug("Start command received.")
-
-    def stop(self):
-        self.started = StartedFlag(False)
-        self.debug("Stop command received.")
-
-    def pause(self):
-        self.paused = StartedFlag(True)
-        self.debug("Paused command received.")
-
-    def resume(self):
-        self.paused = StartedFlag(False)
-        self.debug("Resume command received.")
-
-
-empty_app_factory.cache_product(ControlInterface, ControlInterface)
+            run(_run())
+        else:
+            event_loop.create_task(self.data_feeder.run())
+            event_loop.create_task(self.data_generator.run())
+            event_loop.create_task(self.data_plotter.run())
